change entities to new logic

+++ app/src/main/java/com/avanade/adnd/entities/BattleParticipant.java
@@ -24,8 +24,8 @@ public class BattleParticipant {
     @JoinColumn(name = "char_id")
     private Character character;
     
-    private Short hp;
-    private Short initiative;
+    private Integer hp;
+    private Integer initiative;
     private Boolean playerCharacter;
-
+    
 }
